pyproject.toml
[build-system]
requires = [
  "flit_core>=3.2",
  "fastapi",
  "PyJWT"
]
build-backend = "flit_core.buildapi"

[project]
name = "auth-checker"
version = "2.0.0"
authors = [
  { name="Ryan Semmler", email="rsemmle@ncsu.edu" },
  { name="Luc Sanchez", email="lgsanche@ncsu.edu" },
  { name="Jeremy Gibson", email="jmgibso3@ncsu.edu"},
]
description = "Contains the AuthChecker class for apps using the Auth Service"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastapi==0.110.2",
    "pyjwt==2.8.0",
    "casbin==1.36.0",
    "google-auth==2.19.1",
    "casbin_redis_adapter==1.2.0",
    "casbin_pymongo_adapter==1.1.0",
    "sat-utils==1.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.2",
    "pytest-cov==4.1.0",
    "pytest-mock>=3.14.0, <4.0.0",
    "pytest-env>=1.1.3, <1.2.0",
    "coverage[toml]>=6.2",
    "black>=23.3.0, <24.0.0",
    "ruff==0.1.1",
    "bandit>=1.7.4, <2.0.0",
    "bandit[toml]>=1.7.4, <2.0.0",
    "pre-commit>=2.16.0, <3.0.0",
    "isort>=5.10.1, <6.0.0",
    "pyupgrade>=2.29.1, <3.0.0",
    "flit>=3.8.0, <4.0.0",
    "httpx>=0.27.0, <1.0.0",
    "jupyterlab>=4.1.5, <5.0.0",
]

[project.urls]
"Homepage" = "https://github.com/ncstate-sat/auth_checker"

[tool.flit.module]
name = "auth_checker"

[tool.black]
target-version = ["py310"]
line-length = 100
include = '\.pyi?$'

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "W"]
src = ["auth_checker"]
ignore = ["E501", "F401"]
fix = true

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []


# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "htmlcov",
]

# Same as Black.
line-length = 100

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.bandit]
exclude_dirs = ["tests", "auth_checker/tests"]
skips = ["B101", "B601"]

[tool.pytest_env]
GOOGLE_CLIENT_ID = "google_client_id"
JWT_SECRET = "jwt_secret"
JWT_ALGORITHM = "HS256"
AUTHORIZER = "casbin_authorizer"
CASBIN_AUTHORIZER_POLICY_ADAPTER = "file"
CASBIN_POLICY_FILE = "auth_checker/tests/test_policy.csv"
